focus_guard/requirements.txt
psutil
google-generativeai
plyer
